Skia: Don't run tiles-8888.  Do run tiles_rt-gpu.
Robert and I don't think we're getting much incremental value out of having all
bots run 'tiles' (tiling with no BBH) when they are already running 'tiles_rt'
(tiling with an RTree), so stop running that.

This also seems like a good time to turn on tiled rendering into a GPU backend.
I haven't done much testing beyond my workstations for this mode, so it may be
Fun To Land, but I don't have any specific suspicions.  It seems right to run
it on all bots that run in GPU mode (with the exception of the two Android
problem children).

BUG=skia:3255

Review URL: https://codereview.chromium.org/857353003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293718 0039d316-1c4b-4281-b951-d872f2087c98