Add auto-generation to public waterfall builders.
Update the "chromiumos" master to use the same mechanism as the
"chromeos" master to generate a list of builder names for the status bars.
This solves the immediate problem (and hopefully future problems) of the
bar state being out of sync with the builders.

BUG=chromium:450439
TEST=local
  - Ran script locally, verified correct output.

Review URL: https://codereview.chromium.org/797023008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293737 0039d316-1c4b-4281-b951-d872f2087c98