Add a "check" step on the GN bots..
This change adds a new step by default for all of the bots in this recipe,
which should affect Chromium and V8 but not WebRTC, libyuv, or Mojo.

It might be safe to turn those last three on as well, but let's see if this
causes any problems first.

This change is coded in such a way that it will be easy to turn --check off
bot-by-bot if need be (otherwise we could just hard-code it in the run_gn()
implementation.

I didn't use a recipe 'config' context for this because it felt more awkward and
less clear.

We run this as a separate step rather than 'gn gen --check' because it may take a significant amount of time (at least seconds) and so we're not likely to make this on by default in gn. As a result, if we ran this on the bots then they would be running gen differently than developers mostly will do, which seems like it would be confusing.

BUG=367595, 376000
R=tfarina@chromium.org,brettw@chromium.org, iannucci@chromium.org

Review URL: https://codereview.chromium.org/854763002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293762 0039d316-1c4b-4281-b951-d872f2087c98