bot_update: Infer sensible default for --revision_mapping
Without a default for --revision_mapping bot_update will
set got_revsion to whatever is passed as --revsion, which
is not always useful (e.g. HEAD is not useful value for
got_revsion, neither is the empty string).

With this change if no revision mappying is specified and
the defaults in GOT_REVISION_MAPPINGS do not apply then we
infer a mapping or: { solution0_repo : 'got_revision' }

BUG=450688

Review URL: https://codereview.chromium.org/862913005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293783 0039d316-1c4b-4281-b951-d872f2087c98