Added trace_test to list of gpu tests.
Added the "trace_test" as a test to be run on the bots. Currently this
test will always pass since it will be skipped using this CL:

https://codereview.chromium.org/847773006/

After both of these changes have landed, I will enable it on bots that
support the GL Timer extensions.

R=bajones@chromium.org, vmiura@chromium.org
BUG=None

Review URL: https://codereview.chromium.org/850593002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293786 0039d316-1c4b-4281-b951-d872f2087c98