Remove obsolete sandbox disabling logic from runtest.py.
Previously, we disabled the SUID sandbox by setting CHROME_DEVEL_SANDBOX="" in
the following cases:
- If LSan or TSan were enabled. However, we already pass --no-sandbox in this
case, so setting the env variable to an empty string is unnecessary.
- If the SUID sandbox binary was not present or not configured correctly.
This was useful when SUID sandbox was being rolled out. However, according to
https://code.google.com/p/chromium/wiki/LinuxSUIDSandboxDevelopment
the sandbox binary must now be present on all bots. If it isn't, it makes more
sense to fail than to disable the sandbox. Also, according to the same doc,
setting the env variable to "" is no longer a supported way of disabling the
sandbox.

BUG=245376
R=stip@chromium.org

Review URL: https://codereview.chromium.org/881743004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293832 0039d316-1c4b-4281-b951-d872f2087c98