compile.py support for no-op build checking, and add to Linux clobber builders
This checks ninja's output string for "no work to do". If we decide to add
support for that query to ninja (with a returncode or similar) we can amend
this checking code to be a bit simpler/more robust.

This only adds the check to the Linux clobber builders. We would especially
like it on Windows in the future, but it doesn't work there yet due to
always-build steps building the installer archive. (Mac is also currently
minorly broken.)

R=thakis@chromium.org
BUG=342974,451499

Review URL: https://codereview.chromium.org/883933002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293850 0039d316-1c4b-4281-b951-d872f2087c98