Build just chrome instead of everything.
Building all of the tests increases the exposure to compiler errors,
such as this one which recently hit in gmock:
https://connect.microsoft.com/VisualStudio/feedback/details/1106363/internal-compiler-error-with-analyze-due-to-failure-to-infer-move

While a bug has been filed to get gmock changed to work-around this
compiler error, stopping building things besides Chrome in the
interim seems worthwhile. The gmock bug is:
https://code.google.com/p/googlemock/issues/detail?id=172

And, building just Chrome may be more valuable anyway. Building the
unit tests increases the build times and the noise level, which
makes finding important issues trickier. And, some of the
unique warnings in the tests are intentional bugs.

It's a worthwhile experiment that will get the /analyze builder
running again.

BUG=427616

Review URL: https://codereview.chromium.org/888503002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293853 0039d316-1c4b-4281-b951-d872f2087c98