Add chromium_trybot_legacy recipe
The bots will be switched to it in a separate CL.

This will allow removing the legacy code from chromium_trybot recipe
and make it easier to work with it (e.g. GenTests are currently shared between
two codepaths, and in general they're confusing).

Eventually the chromium_trybot_legacy recipe will be removed,
but we can't do that easily now, e.g. because of linux_chromium_asan_rel which
is in CQ, but we don't have resources to increase number of swarming shards
for linux_chromium_asan_rel_ng so that its cycle time doesn't regress
(it matches main waterfall and so runs more tests, and so takes longer).

BUG=380709

Review URL: https://codereview.chromium.org/880323002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293860 0039d316-1c4b-4281-b951-d872f2087c98