Rename CrWinClang(asan) to CrWinAsan.
The win/clang/asan component bot can't be called CrWinClang(asan/dll) since
that causes path names to be too long for the build (because of
http://crbug.com/399990), so it's called CrWinAsan(dll).
For consistency, rename the existing static asan bot from CrWinClang(asan) to
CrWinAsan.

(Since the test list in the src/ directory also contains this name, there will
be a short window where the bot won't know which tests to run.  Also, the name
change means we'll lose bot history.  Since the bot is currently in great shape,
we think that now is a good time to make this change.)

BUG=448790
TBR=timurrrr@chromium.org

Review URL: https://codereview.chromium.org/887093002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293883 0039d316-1c4b-4281-b951-d872f2087c98