Stop running mojo tests on the GN builders.
Now that we have a chromium.mojo waterfall that runs the tests,
there's no need to run the tests on the GN bots.

Also, remove the the TODO about running the tests ... it's not
clear that we'll ever want to run tests on these bots.

R=jamesr@chromium.org
BUG=

Review URL: https://codereview.chromium.org/902473002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293933 0039d316-1c4b-4281-b951-d872f2087c98