Use goma for NaCl compile in goma canary.
We used to see rename error (errno=17) on Windows when building NaCl
untrusted code. So, we added NO_NACL_GOMA for workaround, and made it
false in some trybots to see if it's fixed. Since we didn't see
errno=17 for long, we'd like to extend this experiment to goma canary
to see if it's really fixed or not.

BUG=359729

Review URL: https://codereview.chromium.org/889323004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293935 0039d316-1c4b-4281-b951-d872f2087c98