Turn off Goma on just the chromium.win waterfall GN bots.
It looks like Goma works fairly reliably on the trybots,
but not the waterfall bots. I suspect there is something
wrong w/ the waterfall bots, and we should migrate them
into the golo anyway, so I'm disabling goma on them for now;
it's much more important that Goma be enabled on the try
bots anyway.

TBR=scottmg@chromium.org
BUG=

Review URL: https://codereview.chromium.org/902443004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293936 0039d316-1c4b-4281-b951-d872f2087c98