Save last close tree message in gatekeeper_ng.
Gives a more reliable way to ensure that tree was *not* closed by a human.
The change is backwards compatible. That is, if "automatic" is found in last
tree message, it's considered to have been set by a human and won't be
overwritten just as before.

R=stip@chromium.org
BUG=N/A

Review URL: https://codereview.chromium.org/896433002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293941 0039d316-1c4b-4281-b951-d872f2087c98