Add step to delete llvm-build directory.
When compiling the clang tool translation_unit, clang is also compiled
and replaces the pre-built binary. In the next run, this locally compiled
binary would be used instead of the pre-built binary, which produces errors
when using goma. So to fix this, we delete the llvm-build directory and the
pre-built binary will be downloaded again in the next run in the
gclient_runhooks step.

BUG=429684

Review URL: https://codereview.chromium.org/902153002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293975 0039d316-1c4b-4281-b951-d872f2087c98