Add __init__.py to, remove pylint disables from scripts/slave/ios.
Need __init__.py to make this a package. Adding this makes pylint complain about the imports not being relative, so those needed to be fixed. Since pylint warnings are being fixed anyways, might as well update for the new pylint in depot_tools and remove as many disables as possible.

However, one pylint warning needs to be active. The import of environment_setup.py is not properly relative (i.e. it should be "from slave.ios import environment_setup.py"). But the specific reason why it's not like this is so that you don't need to have build/scripts in your PYTHONPATH. environment_setup.py is intended to find it for you.

BUG=452193

Review URL: https://codereview.chromium.org/909573002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293983 0039d316-1c4b-4281-b951-d872f2087c98