Don't put revision as part of the local filename.
This avoids the problem of having to clean up. Right now, the
index_pack file archives are cleaned up by relying on the clobber
compile step. This CL also disables the clobber for compilation,
because it does not seem to be really necessary for the Codesearch
bots.

BUG=429684

Review URL: https://codereview.chromium.org/909883005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294006 0039d316-1c4b-4281-b951-d872f2087c98