Change name and config of builder try bot on tryserver.chromium.linux.
In this CL:

 - Change builder name from "bisect_builder" to "linux_full_bisect_builder".
This is so that when there are builders for other platforms, the name wouldn't
conflict. A similar builder on tryserver.chromium.perf is called
"linux_perf_bisect_builder". But if it seems more fitting, this could also be
called "linux_try_builder" or something like that.
 - Change config of the builder so that it archives to the same directory as the
builder on chromium.linux.

The purpose of this is to change the archive location to the place where
auto-bisect expects it, and so that auto-bisect can download builds archived by
either Linux Builder on chromium.linux or the trybot on
tryserver.chromium.linux, from the same place.

BUG=

Review URL: https://codereview.chromium.org/897313003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294026 0039d316-1c4b-4281-b951-d872f2087c98