WebRTC: Make clean script return successful if clean dir is missing.
I thought it would be good to have the script fail for an
incorrectly specified directory, but the out/ dir is missing
on fresh installed slave machines, so better just skip cleaning
with a line printed to stdout instead when the dir is missing.

TESTED=Ran the script locally with an invalid dir argument.

Review URL: https://codereview.chromium.org/921633005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294057 0039d316-1c4b-4281-b951-d872f2087c98