Fix chromium_tests.analyze in the case of compiled_targets without "all".
Before this CL, the compiled_targets passed to chromium_tests.analyze
was ignored and the result of does_patch_require_compile was always
used. This was not a problem in the case where compiled_targets
includes "all", but is a problem in the case compiled_targets does
not include "all".
This CL fixes it and adds tests for chromium_tests.analyze.

TEST=./scripts/slave/unittests/recipe_simulation_test.py test
BUG=430329
R=sky@chromium.org, thakis@chromium.org

Review URL: https://codereview.chromium.org/728653005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294075 0039d316-1c4b-4281-b951-d872f2087c98