Update Python script bootstrapping, Chromite DEPS
Allow 'runit' to intelligently bootstrap Python scripts. If the first
'runit' argument is 'python', it will replace it with the Python
executable. This will allow the Chromite cache to be loaded on Windows.
This should replace the reverted CL:
https://codereview.chromium.org/914303007/

Update the Chromite 'gclient' hook in 'build' to be recursive. This will
cause slave checkouts to properly load the Chromite cache.

Turns on Chromite cache refreshing during script execution. This
is a failsafe to keep slaves working/happy even if the hook configuration
doesn't work. The cache should revert to read-only later if this CL
works as intended.

BUG=chromium:452528
TEST=local
  - Ran locally, verified that it works.

R=dyen@chromium.org, stip@chromium.org

Review URL: https://codereview.chromium.org/926453003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294149 0039d316-1c4b-4281-b951-d872f2087c98