Fixed logcat trying to be uploaded on infra failures.
Currently, the logcat file from appurify is uploaded if the AMP
collect step fails. However, if an infra failure caused the collect
step to fail, the logcat will tried to be uploaded and that step will also
fail. What we want to happen, is have the logcat file uploaded when the step
failure is caused by test failure. This change will at least make it so if
the step status is step.EXCEPTION (infra failure) the logcat file will not
be uploaded.

BUG=

Review URL: https://codereview.chromium.org/976883002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294330 0039d316-1c4b-4281-b951-d872f2087c98