Add a step to update goma canary.
Goma canary buildbot must have a step to update goma to the head
version. We have a few options to achieve that.

1) update run_annotated.py
2) have a new step

Updating run_annotated.py is a bit dangerous since it is shared in all
bots, so we'd like to have a new step to do so.

BUG=452797

Review URL: https://codereview.chromium.org/970793002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294340 0039d316-1c4b-4281-b951-d872f2087c98