Switch the chromium.webkit retina bot to a builder+tester
We seem to be seeing a lot of failures after switching the chromium.webkit WebKit Mac Retina bot to a test-only bot running the chromium recipe.

This change attempts to switch the bot to a builder+tester to see if doing local builds (which is what the bot was doing pre-recipe) fixes things.

TBR=phajdan.jr@chromium.org
BUG=464459
BUG=

Review URL: https://codereview.chromium.org/982883005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294364 0039d316-1c4b-4281-b951-d872f2087c98