V8 Buildbot: Make clusterfuzz issues stop the release process earlier.
Without this change, the release process bot would promote
the candidate ref also with clusterfuzz issues present.
If then at some later point the clusterfuzz issues are
resolved, the roll ref can't be updated to that last
candidate as there's no record of it, leading to a delay
until the next candidate is ready.

After this change, both candidate and roll are on hold
until clusterfuzz issues are resolved.

BUG=443153
TBR=kjellander@chromium.org

Review URL: https://codereview.chromium.org/994513004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294392 0039d316-1c4b-4281-b951-d872f2087c98