WebRTC: Remove got_revision hack.
The chromium.webrtc.fyi waterfall used to override got_revision
since that's the property that's used for displaying the revision
in the builds lists, to make it easier to find the bad WebRTC
revision causing a breakage. It makes things harder when moving to
Git, so let's just remove it.

I changed the expectation revisions slightly to make it more obvious
which one is a Chromium revision and not.

BUG=438170

Review URL: https://codereview.chromium.org/995733002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294399 0039d316-1c4b-4281-b951-d872f2087c98