Do not raise exception when iOS test runner fails to clean up tmp dir
We're flakily seeing OSError: [Errno 66] Directory not empty, which means shutil.rmtree failed to remove something inside the directory we're trying to remove. While this is technically problematic, since we created this directory and should be able to remove it and its contents, there's not much we can do about it. We certainly shouldn't be failing the build when this happens.

TBR=

Review URL: https://codereview.chromium.org/1004933002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294425 0039d316-1c4b-4281-b951-d872f2087c98