WebRTC: Rename revision variables.
I realized that our Chromium builders have been passing a Git hash
to the perf dashboard for a long time (but the WebRTC standalone
builders are passing a SVN/commit position integer).
Because of that, it makes sense to rename the variables to avoid
confusion reading the code.

TBR=machenbach@chromium.org

Review URL: https://codereview.chromium.org/1012843002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294458 0039d316-1c4b-4281-b951-d872f2087c98