Mark as FAILURE when diagnose_goma failed.
diagnose_goma is marked as SUCCESS when it's failed now.
ok_ret makes the step SUCCESS, so it's not what we want to use.

When diagnose_goma failed (exit with non-zero), we'd like to make
the step FAILURE, but since it's not critical, we'd like to proceed
to the next step.

BUG=452797

Review URL: https://codereview.chromium.org/1038583002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294561 0039d316-1c4b-4281-b951-d872f2087c98