WebRTC: Reshuffle MSan slave machines.
Recently, MSan is capable of downloading
prebuilt instrumented dependencies, which eliminates
the hassle of having to do that on each slave VM manually.
Because of that, there's no point keeping the MSan bot
limited to a small number of VMs (to lower the manual maintenance).

Also, the linux_memcheck bot is optional, so it won't hog
slave VMs very often even if it shares the same pool of machines
as the others.

BUG=

Review URL: https://codereview.chromium.org/1031923004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294598 0039d316-1c4b-4281-b951-d872f2087c98