Skipping build url composition when url is given.
This is currently breaking some test runs for the bisect recipes: The
reason is that some build properties required to compose the legacy url
are not present. However, since we are already giving a url to use we
shouldn't be trying to build one in those cases.

R=qyearsley@chromium.org,iannucci@chromium.org

Review URL: https://codereview.chromium.org/1045583002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294617 0039d316-1c4b-4281-b951-d872f2087c98