Skia: remove NoNeon and NoThumb builders.
We're not really interested in these configs anymore.  We're happy with
Thumb-only everywhere it's supported.  As for NEON, we're going from tristate
{Yes, RuntimeCheck, No} to just {Yes, RuntimeCheck}.

Bots like Build-Ubuntu-GCC-Arm7-Debug-Android cover the RuntimeCheck option,
and the Build-Ubuntu-GCC-Arm7-Debug-Android_Neon bots over the Yes option.

While I'm here, use the new "nexus_9" alias for "arm64" for the Nexus 9,
so it fits in with its friends.


BUG=

Review URL: https://codereview.chromium.org/1056963002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294659 0039d316-1c4b-4281-b951-d872f2087c98