Add support for tryservers to builders.pyl/buildbot-tool.
This CL makes it possible for builders.pyl to support tryservers
as well as waterfall bots.

It adds a new 'master_type' field to the builders.pyl format, with
two possible values: "waterfall" and "tryserver". It uses that
value to decide which set of templates to use, from the directories
under scripts/tools/buildbot_tool_templates.

It also adds optional fields to the builders.pyl format for
'buildbucket_bucket' (a string) and 'service_account_file'
(a string). The former should be the bucket assigned to this master for cr-buildbucket.appspot.com; the latter
should point to a filename under '/creds/service_accounts' that
we will use to obtain the OAuth service account credentials for the
slaves.

This CL updates the two existing builders.pyl-based masters to
add the new field, and regenerates their master configs, but as
expected, regenerating the configs did not produce any changes.

BUG=356974
R=luqui@chromium.org, nodir@chromium.org

Review URL: https://codereview.chromium.org/1052433002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294676 0039d316-1c4b-4281-b951-d872f2087c98