Add a recipe to check out gyp and run its tests.
This will be used by the new 'client.gyp' master.

The recipe is a bit more complicated than it needs to be, in that
we tell it to use gclient when we really only need to check out
a single repo, but using gclient will make things more consistent
with 'fetch gyp' and may make this recipe slightly more useful as
a simple template for other small public projects.

The recipe does *not* use a BUILDERS dict or a per-builder set
of tests, since there is no branching in the recipe and since
(I think?) we're gradually moving control flow out of the recipes
and into checkout-side scripts and configuration, and so the
per-builder baselines are less and less useful.

R=iannucci@chromium.org, bradnelson@chromium.org
BUG=472266

Review URL: https://codereview.chromium.org/1060093002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294695 0039d316-1c4b-4281-b951-d872f2087c98