ASan: Do not override strict_memcmp and replace_intrin in runtest.py.
replace_intrin=0 was relevant only on Mac, where we now have default sanitizer
options. This means we no longer need to set it at runtime.
strict_memcmp=0 may still be required on CrOS, where default options are not yet
supported. However, the bug is reported to have been fixed, and this is the only
way to check.

BUG=178677, 162461, 302040
TBR=glider@chromium.org

Review URL: https://codereview.chromium.org/1076483002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294724 0039d316-1c4b-4281-b951-d872f2087c98