Re-reland of 'Make subannotations more robust' (an increasingly dubious claim)
Original: https://codereview.chromium.org/1001183002
Reland: https://codereview.chromium.org/1076643002/

This patch makes the recipe engine parse and re-emit subannotations when allow_subannotations is True.  This means that any of the subannotator's steps which are duplicates of other steps executed will be disambiguated.  When an allow_subannotations step is actually emits annotations, an (end) step is emitted after it is done to have something to hang metadata on (it also conveniently communicates when we re-enter recipes).

BUG=466409
TBR=iannucci@chromium.org

Review URL: https://codereview.chromium.org/1075053002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294761 0039d316-1c4b-4281-b951-d872f2087c98