Update chromium_gn recipe so that it can use the new 'mb' wrapper.
'mb' is a new script in the chromium repo that determines
whether to use GYP or GN to generate the ninja files for a build,
and whether to use GYP or GN for the 'analyze' command.

This CL adds support for mb to the chromium_gn recipe so that we
can incrementally toggle on MB or continue to use the older method
of calling GN directly.

As a result of this change, the 'analyze' step should now work correctly
on the GN bots.

Once we can test MB on a few bots, I plan to update the chromium_trybot
and chromium recipes to be aware of MB, and merge the chromium_gn
recipe into the other recipes.

R=iannucci@chromium.org
TBR=phajdan.jr@chromium.org
BUG=466436, 454413

Review URL: https://codereview.chromium.org/1076993002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294770 0039d316-1c4b-4281-b951-d872f2087c98