CrOS: Update public waterfall to auto-configure.
Update the 'chromiumos' public waterfall to auto-configure from the
pinned Chromite instance.

This change modifies the builder directory names on the master to
conform to a global CrOS standard. When this commit is implemented,
build directories will have to be updated on master to preserve builder
data.

BUG=chromium:456203
TEST=local
  - Started master, confirmed that builder population and slave
    allocation is correct.

Review URL: https://codereview.chromium.org/1068263003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294775 0039d316-1c4b-4281-b951-d872f2087c98