Add a new master.client.gyp and update buildbot_tool to support it.
buildbot_tool's default naming scheme doesn't work that well for
client.X projects or projects whose name is an acronym, so this
patch adds the ability to specify the master_classname in the
builders.pyl file. This patch also fixes a couple of regressions
in the chromium_utils unitttests that had crept in.

R=luqui@chromium.org, iannucci@chromium.org
BUG=472266

Review URL: https://codereview.chromium.org/1067593002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294788 0039d316-1c4b-4281-b951-d872f2087c98