Reenable MB on the Mac and Win waterfall bots.
This CL re-lands part of the MB changes from yesterday, so that
we can incrementally test things and verify that they work, rather
than changing all of the bots at once.

This patch enables MB on the Mac GN, Win GN, and Win GN (dbg) bots.
It does not enable the Mac GN (dbg) bot, and it actually turns off
MB on the mac_chromium_gn_dbg trybot, because the configurations there
are wrong (currently MB will use a shared build, not a static build);
we need to fix the src-side config first before enabling those two.

The changes in GenSteps() themselves reflect the state of the recipe
before I reverted it (I think it was actually correct), and the changes
in GenTests() are differently only to reflect the fact that we're not
yet running mb on the linux trybots.

TBR=phajdan.jr@chromium.org
BUG=454413

Revert "Revert GN -> MB recipe changes"

This reverts commit 40b195c712737e125547f0e0f51398fce49f4f6f.

Review URL: https://codereview.chromium.org/1084283002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294826 0039d316-1c4b-4281-b951-d872f2087c98