Skia: tear down the iOS builders for now.
I'm working on getting NEON to work with our build for iOS, and the
compiler on these build bots is old enough to be getting in the way. 
I can build and run NEON from my desktop just fine, but the
buildbot compiler barfs on the same code.

Stephan is working on bringing up a real iOS tester, which has the
same compiler as my desktop.

BUG=

Review URL: https://codereview.chromium.org/1095663002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294874 0039d316-1c4b-4281-b951-d872f2087c98