xvfb: set dpi to 96 (xvfb defaults to 100).
Tests assume that the DPI is effectively 96. Up until recently, this was
the case because chromium only supported 100% (corresponding to 96 dpi)
or 200% (corresponding to 192 dpi) scaling in its UI.

With https://codereview.chromium.org/1078423002/, we want to change that
to better reflect the hardware that many people have today. Applying
that CL without also changing the DPI results in tests failing because
of off-by-a-few-pixel errors.

BUG=477716,143619

Review URL: https://codereview.chromium.org/1105903002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294979 0039d316-1c4b-4281-b951-d872f2087c98