Lift cf_archive_script into recipe module.
This migrates the contents of the legacy script
cf_archive_script.py into the archive recipe module.

This is the first migration step with a focus on preserving
the old behavior (mostly) and on easy review (by keeping
the old code structure).

In a follow up CL this code should be refactored and
cleaned up.

BUG=478039

Review URL: https://codereview.chromium.org/1091263002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294991 0039d316-1c4b-4281-b951-d872f2087c98