Enable goma for all NaCl builds (retry).
This is the 3rd revival of r294339. The reason of failure is that
archive contains 0 byte object file, so LINK fails.
We thought that object file was not created correctly when archiving.
However, it was not. All object files were not 0 byte. Then, maybe
archiver couldn't read object file in some reason.

We have added a new check in archiving process in NaCl build. We check
the content of archive, and if it contains 0 byte object, we
regenerate archive again (at most 3 times). I've confirmed that the
archiver can create valid archive from the object files that produced
invalid archive. So, I believe the problem is a kind of archiver flake,
and I hope we have workaround in nativeclient revision 10a72113.

BUG=390764

Review URL: https://codereview.chromium.org/1093773003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@294999 0039d316-1c4b-4281-b951-d872f2087c98