Refactoring: Extract static archive configurations.
Define constant top-level sets for excluded files instead
of local lists. The sets are only used for inclusion
checks, hence, no non-determinism problems.

The old comments and todos are kept as they were.

Tiny behavioral change: Instead of a not-implemented
exception, we'd see a KeyError if a file exclusion config
for an unsupported platform was used.

BUG=478039

Review URL: https://codereview.chromium.org/1093413004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295002 0039d316-1c4b-4281-b951-d872f2087c98