Skia: disable link-time code generation on compile bots.
The Windows Debug builders are fast (1-2 minutes) but the Release ones are
still slow (10-15 minutes).  Let's see how much of that is bottlenecked on
link-time code generation, something we only do on Windows.  LTCG is only
relevant for our perf bots.  We don't need it on the compile bots (or the test
bots, but let's leave that for a follow up).

BUG=

Review URL: https://codereview.chromium.org/1111263002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295059 0039d316-1c4b-4281-b951-d872f2087c98