Move code from presubmit module back to run_presubmit
This effectively reverts https://codereview.chromium.org/1080383003/.

Originally, the intent was to create run_presubmit_internal, however, it seems
this is not necessary anymore (we will just add configs for internal projects to
the build_internal). The only change to the original run_presubmit module is the
possibility to request authentication to be used, which is needed to access
internal code review site. Finally, the code duplicated in infra_repo_trybot
will eventually be removed when infra and infra_internal will start using
run_presubmit to run presubmit.

R=akuegel@chromium.org, phajdan.jr@chromium.org
BUG=482262, 478651

Review URL: https://codereview.chromium.org/1112443003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295100 0039d316-1c4b-4281-b951-d872f2087c98