Fix chromium_gn recipe to only run tests that were compiled.
Previously, if the test spec listed a test step, we would try to
run it on a tryserver regardless of what analyze returned (i.e.,
regardless of whether a patch modified any files that affected the
test or not).

This patch corrects that to only run the tests that are affected
by the change, and adds a test for this.

R=jam@chromium.org
BUG=483012

Review URL: https://codereview.chromium.org/1117153002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295103 0039d316-1c4b-4281-b951-d872f2087c98