ClangAndroidToTASan: set TARGET_PLATFORM=android
I previously set it to linux, which made the sizes step fail since it
didn't find the expected binaries.

When setting TARGET_PLATFORM=android it was also necessary to set an
android_config, which pulled in a bunch of other stuff. Hopefully
that's all benign.

BUG=452726

Review URL: https://codereview.chromium.org/1128113003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295156 0039d316-1c4b-4281-b951-d872f2087c98