Add a dependency on 'python' to the chromium gn recipe.
Due to some awkward api structuring, the ScriptTest classes
depend on the caller having declared a dependency on the
'python' module. So, when we try to run script-based tests
in the chromium_gn recipe, suddenly we start needing a
dependency we don't directly need and didn't need before.

The proper way to fix this would probably be to have
chromium.steps.generate_script() be aware of the needed
dependency somehow and inject it, but changing that is
perhaps a bit disruptive. Given that the goal of the
work in the bug below is to merge chromium_gn into chromium,
it's probably easier at the moment to just make chromium_gn
match chromium a little closer.

R=iannucci@chromium.org
BUG=481693

Review URL: https://codereview.chromium.org/1141553005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295334 0039d316-1c4b-4281-b951-d872f2087c98