Changing trigger call to make chromium recipe sync to built revision.
Before this fix, the tester bot would sync to TOT then download the prebuilt
image and unzip it. This ended up running run_benchmark at TOT instead of at
the revision being tested.

With this change we match what the pre-recipes bisect script does so that we
can better make the transition to recipes.

A later change will allow the bisector to force the tester to run telemetry at
TOT.

BUG=484931
R=prasadv,qyearsley

Review URL: https://codereview.chromium.org/1149713003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295338 0039d316-1c4b-4281-b951-d872f2087c98