Adjust master_gen.PopulateBuildmasterConfig() to look at master_site_config.
When master_gen was initially written, we had hoped to be able to delete
master_site_config, and so master.cfg would call
master_gen.PopulateBuildmasterConfig() to define the active_master class
on the fly.
    
However, we realized that we still needed master_site_config.py to exist
for various other scripts and tests, even though the buildbot startup
itself didn't need it. So, we added a template to create the site_config
as well. Unfortunately, the code was now duplicated in two places and
potentially not in sync. In particular, the bug below was caused by
master_site_config being updated, but master_gen not being updated.
    
This CL changes the master.cfg template to actually use the
master_site_config class definition, and simplifies master_gen as
a result, so that the class definitions aren't duplicated.
    
Generally, we don't want master.cfg to import any other modules, because
every module that is imported will *not* be reload on 'make config',
and if that module changes we will actually need to restart the
master. However, the fields that are defined in master_site_config.py,
while pulled from builders.pyl, are generally write-once, defined when
the master is initially created, and shouldn't change over the lifetime
of the master. I guess we'll see if that ends up being true in practice.
    
R=nodir@chromium.org, luqui@chromium.org
BUG=490830

Review URL: https://codereview.chromium.org/1152093002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295381 0039d316-1c4b-4281-b951-d872f2087c98