Set the Mac and Win GN waterfall bots to not close the tree.
Since the Mac and Win GN bots are not yet in the CQ, we
often see tree-closing failures for them. We think it
may be preferable for the bots to still stay on the main
waterfall, but not close the tree, as long as alerts for
the failures still show up in sheriff-o-matic.

R=stip@chromium.org
BUG=486098, 439402

Review URL: https://codereview.chromium.org/1162293007

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295523 0039d316-1c4b-4281-b951-d872f2087c98