Run audio_unittests instead of media_unittests on chromium.gpu.fyi bots. (Re-land #1)
This smaller test suite avoids redundancy with tests already running
on other Chromium bots. A follow-on CL will add a flag verifying that
the tests run correctly on physical hardware.

Stopped adding media_unittests and content_unittests as unconditional
dependencies from the GPU recipe, since on the GPU bots, they're only
run on the FYI waterfall. This changed a lot of expectations due to
the changes to the dummy swarming hashes, and also required an update
to one of the tests in the chromium_trybot recipe.

BUG=416594
TBR=iannucci@chromium.org,sergeyberezin@chromium.org,sergiyb@chromium.org

Review URL: https://codereview.chromium.org/1150183013

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295536 0039d316-1c4b-4281-b951-d872f2087c98