Handle a missing slavebuilddir in master_gen better.
Apparently all of the other builders in builders.pyl-based bots
were explicitly setting the slavebuilddir field. You're not
supposed to need to do this, but if we default to None other
scripts in the system apparently are unhappy, so now we default
to the same (escaped) value that buildbot normally uses if it
isn't set.

TBR=nodir@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1158043005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295543 0039d316-1c4b-4281-b951-d872f2087c98